fix(mcp): wire capture refusal into nested agent
