+++ Apps/CLI/Sources/PeekabooCLI/Commands/Base/Runtime/RuntimeHostResolver+Resolution.swift
@@ -1,8 +1,35 @@
 import Darwin
 import PeekabooAutomationKit
+import PeekabooBridge
 import PeekabooCore
 
 extension RuntimeHostResolver {
+    struct ImplicitRemoteCandidate: Equatable {
+        let socketPath: String
+        let requireReusableDaemon: Bool
+        let requiredHostKind: PeekabooBridgeHostKind?
+        let requiresValidatedHistoricalDaemon: Bool
+    }
+
+    struct RemoteCandidatePlan {
+        let explicitSocket: String?
+        let daemonSocketPath: String
+        let runtimeBuildIdentity: String
+        let buildScopedDaemonSocketPath: String?
+        let historicalBuildScopedDaemonTargets: [DaemonControlTarget]
+        let historicalBuildScopedDaemonSocketPaths: [String]
+        let candidates: [ImplicitRemoteCandidate]
+    }
+
+    struct RemoteCandidateValidation {
+        let reusableDaemonStatus: PeekabooDaemonStatus?
+    }
+
+    enum InitialRoutingDecision: Equatable {
+        case local(snapshotInvalidationRemoteSocketPaths: [String])
+        case remote
+    }
+
     struct Resolution {
         let services: any PeekabooServiceProviding
         let hostDescription: String
