+++ Apps/CLI/Sources/PeekabooCLI/Commands/Base/Runtime/RuntimeHostResolver.swift
@@ -7,32 +7,6 @@ import PeekabooCore
 
 @MainActor
 enum RuntimeHostResolver {
-    struct ImplicitRemoteCandidate: Equatable {
-        let socketPath: String
-        let requireReusableDaemon: Bool
-        let requiredHostKind: PeekabooBridgeHostKind?
-        let requiresValidatedHistoricalDaemon: Bool
-    }
-
-    struct RemoteCandidatePlan {
-        let explicitSocket: String?
-        let daemonSocketPath: String
-        let runtimeBuildIdentity: String
-        let buildScopedDaemonSocketPath: String?
-        let historicalBuildScopedDaemonTargets: [DaemonControlTarget]
-        let historicalBuildScopedDaemonSocketPaths: [String]
-        let candidates: [ImplicitRemoteCandidate]
-    }
-
-    struct RemoteCandidateValidation {
-        let reusableDaemonStatus: PeekabooDaemonStatus?
-    }
-
-    enum InitialRoutingDecision: Equatable {
-        case local(snapshotInvalidationRemoteSocketPaths: [String])
-        case remote
-    }
-
     static func resolveServices(options: CommandRuntimeOptions) async throws -> Resolution {
         let environment = ProcessInfo.processInfo.environment
         let configurationInput = PeekabooAutomation.ConfigurationManager.shared.getConfiguration()?.input
